+++ src/AbstractReportingCloud.php
@@ -87,6 +87,21 @@ abstract class AbstractReportingCloud
     const DEFAULT_DEBUG = false;
 
     /**
+     * Document divider - none
+     */
+    const DOCUMENT_DIVIDER_NONE = 0;
+
+    /**
+     * Document divider - new paragraph
+     */
+    const DOCUMENT_DIVIDER_NEW_PARAGRAPH = 1;
+
+    /**
+     * Document divider - new section
+     */
+    const DOCUMENT_DIVIDER_NEW_SECTION = 2;
+
+    /**
      * Properties
      * -----------------------------------------------------------------------------------------------------------------
      */